requirements.txt
altgraph==0.17.4
et-xmlfile==1.1.0
macholib==1.16.3
openpyxl==3.1.5
packaging==24.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyYAML==6.0.2
setuptools==75.2.0
